Add first stab at a Smpp::Receiver class.
This is a copy of the Smpp::Transceiver class with the minimum
changes I needed for it to bind to the SMSC as a "receiver".

I've added some tests to cover as much of the behaviour that
I think I need, so that I can then strip unnecessary functionality
out of the Receiver class in later commits without risking
breaking the behaviour I need.

Note that the tests are not ideal, but hopefully better than nothing.