Handle multi part messages.
Can now handle 5 bit header information being sent in the body of the message.

To fully support all types of mobile concatenated messages, the ability to handle 8 bit body headers will need to be implemented. At the moment this difficult as we require an example message with a 8 bit header.