Handle GSM escaped characters like euro, curly brackets and tilde.
Because these come with an escape character and then an ASCII character, we substitute it for the appropriate ROMAN-8 coding.

The euro character is a special case, since there is no ROMAN-8 encoding for this character. In this case we simply insert a very unlikely token string, and then replace it after the UTF-8 conversion has occurred.