Handle messages with data_coding of 1
Some messages were found with data_coding of 1.  Although some specs state that this means they are encoded as ASCII, when dealing with O2 SMSCs, bits 2 and 3 of the data_coding are used to indicate encoding.  If they are both set to 0 then this indicates default character set, in our case HP-ROMAN8