Simplified UDH decoding by using a 16-bit integer unpack format for the message_id in the case of the 6-octet UDH.
In this way the @udh array always has the same number of elements whether the UDH is in the 5 or 6 octet form.

We changed the expected values in the assertions checking the decoded UDH values to use hex values so they can be related to the raw hex values above.