Move all the UTF-8 encoding stuff out into an example
The contribution from lazyatom was well intended, but doing a bit too much as far as dealing with encoding the data.
I've added a class method to the DeliverSm PDU to allow the assignment of an encoder to handle this responsibility. An
encoder is anything that responds to encode(data_coding, short_message). See the encoding_test for examples.

This code may make a reappearance in the main library after we come up with a better way to configure the behavior. As
of right now I need to do a new release and decided to just move this code out of the way for now.