Merge branch 'master' of https://github.com/raykrueger/ruby-smpp
I toyed with the idea of rebasing our commits on top of raykrueger/master but decided against it in the end.  Not necessarily for any good reason other than that I didn't want to have to force push our master.

We'll want to refactor the transmitter to remove some of the duplication between that and smpp/base.rb (that was refactored in 7f6c01efd857a202de76c803220aaa0521b67ea3).