Remove pdr_storage hash as it is just leaking
The pdr_storage hash used to be passed in by client implementations if they wished to override this behavior.
It is now more sensible for clients to manage message_id to mt_message_id mappings on their own. This code should have been removed.
As it is now it is simply leaking memory by filling up this hash on every MT.